\name{check.consistent.dimensions.TSmodelconcentrate}
\alias{check.consistent.dimensions.TSmodelconcentrate}
\title{check.consistent.dimensions Specific Methods}
\description{See the generic function description.}
\usage{
    \method{check.consistent.dimensions}{TSmodelconcentrate}(obj1, obj2=NULL)
}
\arguments{
    \item{obj1}{a \code{TSmodelconcentrate} object.}
    \item{obj2}{a matrix of time series or a TSdata object.}
}

\seealso{ \code{\link[tframe]{check.consistent.dimensions}} }

%\keyword{DSE}
\keyword{ts}

