\name{check.residuals.TSdataconcentrate}
\alias{check.residuals.TSdataconcentrate}
\alias{check.residuals.TSdatareconstitute}
\alias{check.residuals.concentrated}
\title{check.residuals Specific Methods}
\description{See the generic function description.}
\usage{
    \method{check.residuals}{TSdataconcentrate}(data, ...)
    \method{check.residuals}{TSdatareconstitute}(data, ...)
    \method{check.residuals}{concentrated}(data, ...)
}
\arguments{
    \item{data}{a TSdataconcentrate object.}
    \item{...}{arguments passed to check.residuals.}
}

\seealso{
  \code{\link{check.residuals}} 
  \code{\link{concentrated.check.residuals}} 
}

%\keyword{DSE}
\keyword{ts}

