\name{checkConsistentDimensions.TSmodelconcentrate}
\alias{checkConsistentDimensions.TSmodelconcentrate}
\title{checkConsistentDimensions Specific Methods}
\description{See the generic function description.}
\usage{
    \method{checkConsistentDimensions}{TSmodelconcentrate}(obj1, obj2=NULL)
}
\arguments{
    \item{obj1}{a \code{TSmodelconcentrate} object.}
    \item{obj2}{a matrix of time series or a TSdata object.}
}

\seealso{ \code{\link[tframe]{checkConsistentDimensions}} }

%\keyword{DSE}
\keyword{ts}

