\name{checkResiduals.TSdataconcentrate}
\alias{checkResiduals.TSdataconcentrate}
\alias{checkResiduals.TSdatareconstitute}
\alias{checkResiduals.concentrated}
\title{checkResiduals Specific Methods}
\description{See the generic function description.}
\usage{
    \method{checkResiduals}{TSdataconcentrate}(data, ...)
    \method{checkResiduals}{TSdatareconstitute}(data, ...)
    \method{checkResiduals}{concentrated}(data, ...)
}
\arguments{
    \item{data}{a TSdataconcentrate object.}
    \item{...}{arguments passed to checkResiduals.}
}

\seealso{
  \code{\link[dse1]{checkResiduals}} 
  \code{\link{concentrated.checkResiduals}} 
}

%\keyword{DSE}
\keyword{ts}

