\name{concentrated.check.residuals}
\alias{concentrated.check.residuals}
\title{Check Residuals of Concentrated Data}
\description{
    The TSdataconcentrate is passed to check.residuals as simple TSdata
    in the reduced dimension space (not expanded).
}
\usage{
    concentrated.check.residuals(data, ...)
}
\arguments{
    \item{data}{a TSdataconcentrate object.}
    \item{...}{arguments passed to check.residuals.}
}
\value{x}

\seealso{
    \code{\link[dse1]{check.residuals}}
    \code{\link{concentrate}}
}

%\keyword{DSE}
\keyword{ts}

