\name{concentrated.checkResiduals}
\alias{concentrated.checkResiduals}
\title{Check Residuals of Concentrated Data}
\description{
    The TSdataconcentrate is passed to checkResiduals as simple TSdata
    in the reduced dimension space (not expanded).
}
\usage{
    concentrated.checkResiduals(data, ...)
}
\arguments{
    \item{data}{a TSdataconcentrate object.}
    \item{...}{arguments passed to checkResiduals.}
}
\value{x}

\seealso{
    \code{\link[dse1]{checkResiduals}}
    \code{\link{concentrate}}
}

%\keyword{DSE}
\keyword{ts}

