\name{end.TScanonical.prediction}
\alias{end.TScanonical.prediction}
\alias{start.TScanonical.prediction}
\alias{frequency.TScanonical.prediction}
\alias{periods.TScanonical.prediction}

\title{Specific Methods for TScanonical.prediction}
\description{See the generic function description.}

\usage{
   \method{end}{TScanonical.prediction}(x)
   \method{start}{TScanonical.prediction}(x)
   \method{periods}{TScanonical.prediction}(x)
   \method{frequency}{TScanonical.prediction}(x)
}
\arguments{
    \item{x}{An object containing TSdata.}
}
\value{Depends.}
\seealso{
   \code{\link{end}}
   \code{\link{start}}
   \code{\link[tframe]{periods}}
   \code{\link{frequency}}
}
%\keyword{DSE}
\keyword{ts}

