\name{end.TScanonicalPrediction}
\alias{end.TScanonicalPrediction}
\alias{start.TScanonicalPrediction}
\alias{frequency.TScanonicalPrediction}
\alias{periods.TScanonicalPrediction}

\title{Specific Methods for TScanonicalPrediction}
\description{See the generic function description.}

\usage{
   \method{end}{TScanonicalPrediction}(x, ...)
   \method{start}{TScanonicalPrediction}(x, ...)
   \method{periods}{TScanonicalPrediction}(x)
   \method{frequency}{TScanonicalPrediction}(x, ...)
}
\arguments{
    \item{x}{An object containing TSdata.}
    \item{...}{(further arguments, currently disregarded).}
}
\value{Depends.}
\seealso{
   \code{\link{end}}
   \code{\link{start}}
   \code{\link[tframe]{periods}}
   \code{\link{frequency}}
}
%\keyword{DSE}
\keyword{ts}

