\name{input.dimension.TSmodelconcentrate}
\alias{input.dimension.TSmodelconcentrate}
\alias{output.dimension.TSmodelconcentrate}

\title{Specific Methods for input/output.dimension}
\description{See the generic function description.}
\usage{
    \method{input.dimension}{TSmodelconcentrate}(x)
    \method{output.dimension}{TSmodelconcentrate}(x)
}
\arguments{
    \item{x}{a TSdata or TSmodelconcentrate object.}
}

\seealso{
    \code{\link{input.dimension}}
    \code{\link{output.dimension}}
}

%\keyword{DSE}
\keyword{ts}

