\name{nseriesInput.TSmodelconcentrate}
\alias{nseriesInput.TSmodelconcentrate}
\alias{nseriesOutput.TSmodelconcentrate}

\title{Specific Methods for input/nseriesOutput}
\description{See the generic function description.}
\usage{
    \method{nseriesInput}{TSmodelconcentrate}(x)
    \method{nseriesOutput}{TSmodelconcentrate}(x)
}
\arguments{
    \item{x}{a TSdata or TSmodelconcentrate object.}
}

\seealso{
    \code{\link[dse1]{nseriesInput}}
    \code{\link[dse2]{nseriesOutput}}
}

\concept{DSE}
\keyword{ts}

