\name{percent.change.TScanonical.prediction}
\alias{percent.change.TScanonical.prediction}

\title{Specific Methods for percent.change}
\description{See the generic function description.}

\usage{
   \method{percent.change}{TScanonical.prediction}(obj,
        base=NULL, lag=1, cumulate=FALSE, e=FALSE)}
\arguments{
    \item{obj}{see the generic function.}
    \item{e}{see the generic function.}
    \item{base}{see the generic function.}
    \item{lag}{see the generic function.}
    \item{cumulate}{see the generic function.}
}

\seealso{ \code{\link{percent.change}} }
%\keyword{DSE}
\keyword{ts}

