\name{percentChange.TScanonicalPrediction}
\alias{percentChange.TScanonicalPrediction}

\title{Specific Methods for percentChange}
\description{See the generic function description.}

\usage{
   \method{percentChange}{TScanonicalPrediction}(obj,
        base=NULL, lag=1, cumulate=FALSE, e=FALSE, ...)
}
\arguments{
    \item{obj}{see the generic function.}
    \item{e}{see the generic function.}
    \item{base}{see the generic function.}
    \item{lag}{see the generic function.}
    \item{cumulate}{see the generic function.}
    \item{...}{arguments passed to other methods.}
}

\seealso{\code{\link[tframe]{percentChange}} }
%\keyword{DSE}
\keyword{ts}

