\name{print.concentrate}
\alias{print.concentrate}

\title{Print Specific Methods}
\description{See the generic function description.}
\usage{
    \method{print}{concentrate}(x, ...)
}
\arguments{
    \item{x}{a concentrate object.}
    \item{...}{arguments to be passed to other methods.}
}
\seealso{ \code{\link{print}} }

\concept{DSE}
\keyword{ts}

