\name{select.series.concentrate}
\alias{select.series.concentrate}
\title{Specific Methods for select.series}
\description{See the generic function description.}
\usage{
    \method{select.series}{concentrate}(x,
         series = seq(nrow(concentrator(x)$proj)))
}
\arguments{
    \item{x}{a concentrate data object.}
    \item{series}{vector of strings or integers indicating series to select.}
}
\seealso{ \code{\link[dse1]{select.series}} }

%\keyword{DSE}
\keyword{ts}

