\name{selectSeries.concentrate}
\alias{selectSeries.concentrate}
\title{Specific Methods for selectSeries}
\description{See the generic function description.}
\usage{
    \method{selectSeries}{concentrate}(x,
         series = seq(nrow(concentrator(x)$proj)))
}
\arguments{
    \item{x}{a concentrate data object.}
    \item{series}{vector of strings or integers indicating series to select.}
}
\seealso{\code{\link[dse1]{selectSeries}} }

\concept{DSE}
\keyword{ts}

