\name{settf.concentrate}
\alias{settf.concentrate}

\title{Construct a Tframed Object}
\description{Set the tframe of an object.}

\usage{
    \method{settf}{concentrate}(value, x)
}

\arguments{
    \item{x}{a (tframed) concentrate object or a concentrate object to be tframed.}
    \item{value}{a tframe attribute to be applied to x.}
}
\value{A tframed object.}
\details{See the generic.
}
\seealso{
    \code{\link{tframe}}
}

\keyword{programming}
\keyword{utilities}
\keyword{internal}
