\name{tfplot.concentrated}
\alias{tfplot.concentrated}
\alias{tfplot.TScanonical.prediction}
\alias{tfplot.TSdataconcentrate}
\alias{tfplot.TSdatareconstitute}
\title{tfplot Specific Methods}
\description{See the generic function description.}
\usage{
	tfplot(...)
}

\seealso{ \code{\link{tfplot}} }

\keyword{DSE}
\keyword{time series}

