\name{tfprint.concentrate}
\alias{tfprint.concentrate}

\title{Tfprint Specific Methods}
\description{See the generic function description.}
\usage{
    \method{tfprint}{concentrate}(x, ...)
}
\arguments{
    \item{x}{a concentrate object to print.}
    \item{...}{arguments to be passed to other methods.}
}

\seealso{ \code{\link[tframe]{tfprint}} }

\concept{DSE}
\keyword{ts}

