\name{tfwindow.concentrate}
\alias{tfwindow.concentrate}

\title{tfwindow Specific Methods}
\description{See the generic function description.}
\usage{
    \method{tfwindow}{concentrate}(x, ...)
}
\arguments{
    \item{x}{a concentrate object to trucate.}
    \item{...}{arguments to be passed to other methods.}
}

\seealso{ \code{\link{tfwindow}} }

%\keyword{DSE}
\keyword{ts}

