\name{TSdata.TSdataconcentrate}
\alias{TSdata.TSdataconcentrate}

\title{TSdata Specific Methods}
\description{See the generic function description.}
\usage{
    \method{TSdata}{TSdataconcentrate}(d, conc=NULL, names=NULL)
}
\arguments{
    \item{d}{a TSdataconcentrate object from which to get TSdata.}
    \item{conc}{object containing the concentrator (projection) matrix 
       used for the reduction}
    \item{names}{series names for the result.}
}

\seealso{ \code{\link[dse1]{TSdata}} }

%\keyword{DSE}
\keyword{ts}


