\name{check.consistent.dimensions.TSmodelconcentrate}
\alias{check.consistent.dimensions.TSmodelconcentrate}
\title{check.consistent.dimensions Specific Methods}
\description{See the generic function description.}
\usage{
    \method{check.consistent.dimensions}{TSmodelconcentrate}(model, data=NULL)
}
\arguments{
    \item{model}{a \code{TSmodelconcentrate} object.}
    \item{data}{a matrix of time series or a TSdata object.}
}

\seealso{ \code{\link{check.consistent.dimensions}} }

%\keyword{DSE}
\keyword{ts}

