\name{check.consistent.dimensions.TSmodelconcentrate}
\alias{check.consistent.dimensions.TSmodelconcentrate}
\title{check.consistent.dimensions Specific Methods}
\description{See the generic function description.}
\usage{
	check.consistent.dimensions(...)
}

\seealso{ \code{\link{check.consistent.dimensions}} }

\keyword{DSE}
\keyword{time series}

