\name{check.residuals.TSdataconcentrate}
\alias{check.residuals.TSdataconcentrate}
\alias{check.residuals.TSdatareconstitute}
\alias{check.residuals.concentrated}
\title{check.residuals Specific Methods}
\description{See the generic function description.}
\usage{
	check.residuals(...)
}

\seealso{ \code{\link{check.residuals}} }

\keyword{DSE}
\keyword{time series}

