\name{input.dimension.TSmodelconcentrate}
\alias{input.dimension.TSmodelconcentrate}
\alias{output.dimension.TSmodelconcentrate}

\title{Specific Methods for input/output.dimension}
\description{See the generic function description.}
\usage{
	input.dimension(m)
	output.dimension(m)
}

\seealso{
	\code{\link{input.dimension}}
	\code{\link{output.dimension}}
}

\keyword{DSE}
\keyword{ts}

