\name{TSdata.TSdataconcentrate}
\alias{TSdata.TSdataconcentrate}

\title{TSdata Specific Methods}
\description{See the generic function description.}
\usage{
    \method{TSdata}{TSdataconcentrate}(d, conc=NULL, names=NULL)
}
\arguments{
    \item{d}{a TSdataconcentrate object from which to get TSdata.}
}

\seealso{ \code{\link{TSdata}} }

%\keyword{DSE}
\keyword{ts}


