\name{TSdata.TSdataconcentrate}
\alias{TSdata.TSdataconcentrate}

\title{TSdata Specific Methods}
\description{See the generic function description.}
\usage{
    \method{TSdata}{TSdataconcentrate}(data, names=NULL, ...)
}
\arguments{
    \item{data}{a TSdataconcentrate object from which to get TSdata.}
    \item{names}{series names for the result.}
    \item{...}{arguments to be passed to other methods.}
}
\details{Uses reconstitute to build TSdata.
}
\seealso{
   \code{\link{reconstitute}} 
   \code{\link[dse1]{TSdata}} 
}

\concept{DSE}
\keyword{ts}


